Fixes to install.sh

+++ install.sh
@@ -7,9 +7,9 @@ curl -Lo lazygit.tar.gz "https://github.com/jesseduffield/lazygit/releases/lates
 tar xf lazygit.tar.gz lazygit
 sudo install lazygit /usr/local/bin
 
+sudo apt update -y
 sudo apt install -y yadm
 sudo add-apt-repository -y ppa:maveonair/helix-editor
-sudo apt update -y
 sudo apt install -y helix
 
 wget https://github.com/zellij-org/zellij/releases/latest/download/zellij-x86_64-unknown-linux-musl.tar.gz
@@ -17,6 +17,5 @@ tar -xvf zellij*.tar.gz
 chmod +x zellij
 sudo mv -f zellij /usr/bin/
 
-mv .bashrc .backup_bashrc
+[ -f .bashrc ] && mv .bashrc .backup_bashrc
 yadm clone -f https://github.com/neerajanil/dotfiles
-
